2024.12.22 新增特殊的log书写方式
