+++ test/test_log_old.py
@@ -1,0 +1,14 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+# File          : test_log
+# Author        : Sun YiFan-Movoid
+# Time          : 2024/8/22 0:44
+# Description   : 
+"""
+from movoid_log import Log
+
+
+class TestLogOld:
+    def test_log_old(self):
+        log = Log()
